remove old relayer
