--- fuel-relayer-new/Cargo.toml
@@ -1,54 +1,0 @@
-[package]
-name = "fuel-relayer-new"
-version = "0.10.1"
-authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2021"
-homepage = "https://fuel.network/"
-keywords = ["blockchain", "fuel", "fuel-vm"]
-license = "BUSL-1.1"
-repository = "https://github.com/FuelLabs/fuel-core"
-description = "Fuel Relayer"
-
-[dependencies]
-anyhow = "1.0"
-async-trait = "0.1"
-bytes = "1.1"
-chrono = "0.4"
-env_logger = "0.9"
-ethers-contract = { version = "0.17", default-features = false, features = [
-    "abigen",
-] }
-ethers-core = { version = "0.17", default-features = false }
-ethers-middleware = { version = "0.17", default-features = false }
-ethers-providers = { version = "0.17", default-features = false, features = [
-    "ws",
-    "rustls",
-] }
-ethers-signers = { version = "0.17", default-features = false }
-features = "0.10"
-fuel-core-interfaces = { path = "../fuel-core-interfaces", package = "fuel-core-interfaces", version = "0.10.1" }
-futures = "0.3"
-hex = "0.4"
-once_cell = "1.4"
-parking_lot = "0.12"
-serde = "1.0"
-serde_json = "1.0"
-sha3 = "0.10"
-thiserror = "1.0"
-tokio = { version = "1.20", features = ["macros"]}
-tracing = "0.1"
-tracing-subscriber = "0.3.9"
-url = "2.2"
-
-[dev-dependencies]
-fuel-relayer-new ={ path = ".", features = ["test-helpers"]}
-fuel-core-interfaces = { path = "../fuel-core-interfaces", package = "fuel-core-interfaces", features = [
-    "test-helpers",
-] }
-rand = "0.8"
-tracing-test = "0.2"
-test-case = "2.2"
-
-[features]
-test-helpers = ["fuel-core-interfaces/test-helpers"]
-
