--- fuel-relayer-new/src/lib.rs
@@ -1,25 +1,0 @@
-//! # Fuel Relayer
-
-#![forbid(unsafe_code)]
-#![warn(missing_docs)]
-
-pub(crate) mod abi;
-pub(crate) mod config;
-pub(crate) mod log;
-
-mod relayer;
-
-#[cfg(any(test, feature = "test-helpers"))]
-pub mod mock_db;
-#[cfg(any(test, feature = "test-helpers"))]
-pub mod test_helpers;
-
-#[cfg(any(test, feature = "test-helpers"))]
-pub use abi::*;
-
-pub use config::Config;
-pub use ethers_core::types::{
-    H160,
-    H256,
-};
-pub use relayer::RelayerHandle;
