+++ fuel-relayer/Cargo.toml
@@ -41,11 +41,13 @@ tracing-subscriber = "0.3.9"
 url = "2.2"
 
 [dev-dependencies]
+fuel-relayer = { path = ".", features = ["test-helpers"]}
 fuel-core-interfaces = { path = "../fuel-core-interfaces", package = "fuel-core-interfaces", features = [
     "test-helpers",
 ] }
 rand = "0.8"
 tracing-test = "0.2"
+test-case = "2.2"
 
 [features]
 test-helpers = ["fuel-core-interfaces/test-helpers"]
